:art: 抽离storeDecorator
